+++ mktorrent.h
@@ -30,30 +30,30 @@ struct flist_s;
 typedef struct flist_s flist_t;
 struct flist_s {
 	char *path;
-	off_t size;
+	unsigned long size;
 	flist_t *next;
 };
 
 typedef struct {
 	/* options */
-	size_t piece_length;		/* piece length */
-	llist_t *announce_list;		/* announce URLs */
-	char *comment;			/* optional comment */
-	const char *torrent_name;	/* name of torrent (name of directory) */
-	char *metainfo_file_path;	/* absolute path to the metainfo file */
-	slist_t *web_seed_list;		/* web seed URLs */
-	int target_is_directory;	/* target is a directory */
-	int no_creation_date;		/* don't write the creation date */
-	int private;			/* set the private flag */
-	int verbose;			/* be verbose */
+	unsigned int piece_length; /* piece length */
+	llist_t *announce_list;    /* announce URLs */
+	char *comment;             /* optional comment */
+	const char *torrent_name;  /* name of torrent (name of directory) */
+	char *metainfo_file_path;  /* absolute path to the metainfo file */
+	slist_t *web_seed_list;    /* web seed URLs */
+	int target_is_directory;   /* target is a directory */
+	int no_creation_date;      /* don't write the creation date */
+	int private;               /* set the private flag */
+	int verbose;               /* be verbose */
 #ifdef USE_PTHREADS
-	unsigned int threads;		/* number of threads used for hashing */
+	unsigned int threads;      /* number of threads used for hashing */
 #endif
 
 	/* information calculated by read_dir() */
-	unsigned long long size;	/* combined size of all files */
-	flist_t *file_list;		/* list of files and their sizes */
-	unsigned int pieces;		/* number of pieces */
+	unsigned long size;        /* combined size of all files */
+	flist_t *file_list;        /* list of files and their sizes */
+	unsigned int pieces;       /* number of pieces */
 } metafile_t;
 
 #endif /* _MKTORRENT_H */
